Create specific watcher instead of using a path provider
