+++ src/extension.ts
@@ -4,7 +4,7 @@
 import * as fse from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
-import { commands, Disposable, Extension, ExtensionContext, extensions, FileSystemWatcher, languages, Uri, window, workspace } from 'vscode';
+import { commands, Disposable, Extension, ExtensionContext, extensions, languages, window } from 'vscode';
 import { dispose as disposeTelemetryWrapper, initializeFromJsonFile, instrumentOperation } from 'vscode-extension-telemetry-wrapper';
 import { testCodeLensProvider } from './codeLensProvider';
 import { debugTests, runTests } from './commands/executeTests';
@@ -17,7 +17,7 @@ import { TestTreeNode } from './explorer/TestTreeNode';
 import { logger } from './logger/logger';
 import { ITestItem } from './protocols';
 import { runnerExecutor } from './runners/runnerExecutor';
-import { testPathProvider } from './testPathProvider';
+import { testFileWatcher } from './testFileWatcher';
 import { testReportProvider } from './testReportProvider';
 import { testResultManager } from './testResultManager';
 import { testStatusBarProvider } from './testStatusBarProvider';
@@ -38,28 +38,7 @@ async function doActivate(_operationId: string, context: ExtensionContext): Prom
         throw new Error('Could not find Java home.');
     }
 
-    await testPathProvider.initialize();
-    const watcher: FileSystemWatcher = workspace.createFileSystemWatcher('**/*.{[jJ][aA][vV][aA]}');
-    watcher.onDidChange((uri: Uri) => {
-        const node: TestTreeNode | undefined = explorerNodeManager.getNode(uri.fsPath);
-        if (node) {
-            testExplorer.refresh(node);
-        }
-    }, null, context.subscriptions);
-
-    watcher.onDidDelete((uri: Uri) => {
-        const node: TestTreeNode | undefined = explorerNodeManager.getNode(uri.fsPath);
-        if (node) {
-            explorerNodeManager.removeNode(uri.fsPath);
-            testExplorer.refresh();
-        }
-    }, null, context.subscriptions);
-
-    watcher.onDidCreate((uri: Uri) => {
-        if (testPathProvider.isInTestPaths(uri.fsPath)) {
-            testExplorer.refresh();
-        }
-    }, null, context.subscriptions);
+    await testFileWatcher.initialize(context);
 
     testExplorer.initialize(context);
     runnerExecutor.initialize(javaHome, context);
@@ -76,8 +55,6 @@ async function doActivate(_operationId: string, context: ExtensionContext): Prom
         testResultManager,
         testReportProvider,
         logger,
-        testPathProvider,
-        watcher,
         languages.registerCodeLensProvider({ scheme: 'file', language: 'java' }, testCodeLensProvider),
         instrumentAndRegisterCommand(JavaTestRunnerCommands.OPEN_DOCUMENT_FOR_NODE, async (node: TestTreeNode) => await openTextDocumentForNode(node)),
         instrumentAndRegisterCommand(JavaTestRunnerCommands.REFRESH_EXPLORER, (node: TestTreeNode) => testExplorer.refresh(node)),
